Rewrite the backend in Rust (axum + sqlx)
Full port of the P1 + P2 TypeScript backend to Rust. Same schema (migrations
reused; numeric -> double precision for the display-only rate columns), same
endpoints, same behaviour, verified against a ported integration suite
(49 tests, tower::oneshot against the real Router + Postgres).

- axum 0.8 routing; Session extractor (FromRequestParts); Body<T> JSON
  extractor rendering VALIDATION on malformed input
- sqlx 0.8 runtime queries (no compile-time DB), FromRow structs; migrations
  embedded via sqlx::migrate!
- ApiError implements IntoResponse -> { code, message, retryable } with the
  §02 status map; codes serialize SCREAMING_SNAKE
- contract/ = hand-maintained serde mirror of Kimana_frontend/src/api/types
  (no shared compiler; fidelity is a review discipline — see plan §00)
- domain modules: onboarding (+ kyb stub), dashboard, fx, recipients, quote,
  transfers (state_machine + engine), ledger (post_entry under a row lock)
- transfer progression simulated via tokio::spawn; TRANSFER_AUTO_ADVANCE_MS
  (-1 disables for tests, which call the engine directly)
- bins: kimana-backend (serve), migrate, seed
- integration/live-api-client.ts extended to cover P2; wire contract unchanged
- docs/backend-plan.md §00/§01/§06/§07 updated for the language change

TypeScript tree, package.json, node_modules removed.

+++ migrations/001_initial_slice.sql
@@ -58,7 +58,7 @@ create table onboarding_principals (
   position              int not null,
   full_name             text not null,
   role                  text not null check (role in ('director', 'beneficial_owner', 'both')),
-  ownership_percentage  numeric,
+  ownership_percentage  double precision,
   date_of_birth         date,
   bvn                   text,
   nin                   text
